feat: HotNailList.tsx component 분리 #17
- HotNailList.tsx 내부에서 사용된 Card를 별도의 NailInfoCard component로
  분리
- Back-end API 연동 작업을 위한 axios install
